Add: Confirmação ao deletar alunos ou cursos.

+++ login/home/adm/deletar.php
@@ -9,26 +9,50 @@
         content="Edição de dados">
     </head>
 
-    <body>
-    <?php
-        session_start();
-        include "../../../const.php";
-        extract($_GET);
+    <body>       
+        <?php
+            session_start();
+            include "../../../const.php";
+            extract($_GET);
+            extract($_POST);
 
-        if(isset($idCurso)){
+            if(isset($idCurso)){
+                echo "
+                    <form action='' method='POST'>
+                        Confirma exclusão do curso $nomeCurso?
+                        <input type='submit' name='sim' value='Sim'>
+                        <input type='submit' name='nao' value='Não'>
+                    ";
 
-            $consulta = "DELETE FROM `curso` WHERE idCurso = $idCurso";
-            banco($server, $user, $password, $db, $consulta);
-            header('Location: relatorioCursos.php');
+                    if(isset($sim)){
+                        $consulta = "DELETE FROM `curso` WHERE idCurso = $idCurso";
+                        banco($server, $user, $password, $db, $consulta);
+                        header('Location: relatorioCursos.php');
+                    }elseif(isset($nao)){
+                        header('Location: relatorioCursos.php');
+                    }
+                
 
-        }elseif(isset($matricula)){
+            }elseif(isset($matricula)){
+                echo "
+                    <form action='' method='POST'>
+                        Confirma exclusão do aluno $nome?
+                        <input type='submit' name='sim' value='Sim'>
+                        <input type='submit' name='nao' value='Não'>
+                    ";
 
-            $consulta = "DELETE FROM `aluno` WHERE matricula = $matricula";
-            banco($server, $user, $password, $db, $consulta);
-            header('Location: relatorioAlunos.php');
+                    if(isset($sim)){
+                        $consulta = "DELETE FROM `aluno` WHERE matricula = $matricula";
+                        banco($server, $user, $password, $db, $consulta);
+                        header('Location: relatorioAlunos.php');
+                    }elseif(isset($nao)){
+                        header('Location: relatorioAlunos.php');
+                    }
 
-        }
+                
 
-    ?>
+            }
+
+        ?>
     </body>
 </html>
